requirements.txt
# Video processing
moviepy
opencv-python
imageio
imageio-ffmpeg

# Image processing
Pillow
numpy

# YAML processing
PyYAML

# Text rendering and fonts
arabic-reshaper
python-bidi

# Rich terminal output
rich

# Audio processing
pydub

# HTTP requests for AI background generation
requests

# Color manipulation
colorthief

# Progress bars
tqdm

# Date handling
python-dateutil

# ComfyUI integration
websocket-client

# Optional: TikTok upload automation
# tiktok-uploader  # Uncomment if needed